# public/api/openapi.yaml
openapi: 3.1.0
info:
  title: Phototracker API
  version: 0.1.0
  license:
    name: Proprietary
  description: |
    API for tracking film photographs — cameras, lenses, film stocks, development profiles, film holders, rolls, exposures, and reference images.
servers:
  - url: https://darkcloth.zone
    description: Primary production hostname.
  - url: https://phototracker.graha.ms
    description: Legacy/staging hostname during the transition.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Roll:
      type: object
      required: [id, user_id, name, status, push_pull_stops, created_at]
      properties:
        id:
          type: string
          description: ULID
          example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
        user_id:
          type: string
          example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
        film_id:
          type:
            - string
            - "null"
          example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
        roll_format:
          type:
            - string
            - "null"
          enum:
            - 35mm
            - 120
            - 220
            - 127
            - 620
            - null
          description: Roll media format. Null means the roll is format-agnostic.
        name:
          type: string
          example: "Portra 400 #1"
        loaded_at:
          type:
            - string
            - "null"
          format: date-time
          example: "2026-04-29T20:26:05Z"
        finished_at:
          type:
            - string
            - "null"
          format: date-time
          description: Timestamp when the roll was marked ready for processing.
          example: null
        status:
          $ref: "#/components/schemas/RollStatus"
        push_pull_stops:
          type: integer
          description: Push/pull adjustment in stops. Valid values range from -3 to 3.
          enum: [-3, -2, -1, 0, 1, 2, 3]
          example: 0
        processed_at:
          type:
            - string
            - "null"
          format: date-time
          description: Canonical processing completion timestamp. `developed_at` remains a compatibility alias.
          example: null
        developed_at:
          type:
            - string
            - "null"
          format: date-time
          deprecated: true
          description: Compatibility alias for `processed_at` returned for older callers.
          example: null
        development_profile_id:
          type:
            - string
            - "null"
          description: Development profile used when the roll was processed.
          example: "01ARZ3NDEKTSV4RRFFQ69G5FAV"
        development_notes:
          type:
            - string
            - "null"
          description: Free-form notes recorded when the roll was processed.
          example: "Processed in DDX"
        created_at:
          type: string
          format: date-time
          example: "2026-04-29T20:26:05Z"
    RollStatus:
      type: string
      enum:
        - unexposed
        - exposing
        - finished
        - processed
        - developed
      description: Canonical roll lifecycle status. `processed` is the canonical final state; `developed` remains a compatibility alias.
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
...
    limit:
      name: limit
      in: query
      schema:
        type: integer
        default: 50
        minimum: 1
        maximum: 200
    offset:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0
security:
  - bearerAuth: []
paths:
...
  /api/film/rolls:
    get:
      operationId: listRolls
      summary: List rolls
      parameters:
        - $ref: "#/components/parameters/limit"
        - $ref: "#/components/parameters/offset"
        - name: film_id
          in: query
          schema:
            type: string
        - name: roll_format
          in: query
          schema:
            type: string
            enum:
              - 35mm
              - 120
              - 220
              - 127
              - 620
          description: Filter rolls by media format.
      responses:
        "200":
          description: Roll list
          content:
            application/json:
              schema:
                type: object
                required: [items, total]
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Roll"
                  total:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createRoll
      summary: Start a roll
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                film_id:
                  type:
                    - string
                    - "null"
                roll_format:
                  type:
                    - string
                    - "null"
                  enum:
                    - 35mm
                    - 120
                    - 220
                    - 127
                    - 620
                    - null
                  description: Roll media format. Null means the roll is
                    format-agnostic.
                loaded_at:
                  type:
                    - string
                    - "null"
                  format: date-time
                finished_at:
                  type:
                    - string
                    - "null"
                  format: date-time
                processed_at:
                  type:
                    - string
                    - "null"
                  format: date-time
                developed_at:
                  type:
                    - string
                    - "null"
                  format: date-time
                  deprecated: true
                  description: Compatibility alias for `processed_at`.
                development_profile_id:
                  type:
                    - string
                    - "null"
                  description: Development profile used when the roll is processed.
                development_notes:
                  type:
                    - string
                    - "null"
                  description: Free-form notes recorded when the roll is processed.
                push_pull_stops:
                  type: integer
                  enum: [-3, -2, -1, 0, 1, 2, 3]
                  default: 0
      responses:
        "201":
          description: Roll created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roll"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/film/rolls/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      operationId: getRoll
      summary: Get a roll
      responses:
        "200":
          description: Roll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roll"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateRoll
      summary: Update a roll
      description: >
        Update roll metadata or lifecycle timestamps. `developed_at` remains a
        compatibility alias for `processed_at`. When lifecycle timestamps are
        changed, the roll status is updated to the matching current lifecycle
        state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                film_id:
                  type:
                    - string
                    - "null"
                roll_format:
                  type:
                    - string
                    - "null"
                  enum:
                    - 35mm
                    - 120
                    - 220
                    - 127
                    - 620
                    - null
                  description: Roll media format. Null clears the stored
                    format.
                loaded_at:
                  type:
                    - string
                    - "null"
                  format: date-time
                finished_at:
                  type:
                    - string
                    - "null"
                  format: date-time
                processed_at:
                  type:
                    - string
                    - "null"
                  format: date-time
                developed_at:
                  type:
                    - string
                    - "null"
                  format: date-time
                  deprecated: true
                  description: Compatibility alias for `processed_at`.
                development_profile_id:
                  type:
                    - string
                    - "null"
                  description: Development profile used when the roll is processed. The profile must belong to the current user and match the roll's film stock when the roll has a film stock.
                development_notes:
                  type:
                    - string
                    - "null"
                  description: Free-form notes recorded when the roll is processed.
                push_pull_stops:
                  type: integer
                  enum: [-3, -2, -1, 0, 1, 2, 3]
      responses:
        "200":
          description: Updated roll
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roll"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteRoll
      summary: Delete a roll
      responses:
        "204":
          description: Deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /api/film/rolls/{id}/finish:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      operationId: finishRoll
      summary: Mark a roll finished
      description: Marks the roll ready for processing. The timestamp defaults to now when omitted.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                finished_at:
                  type:
                    - string
                    - "null"
                  format: date-time
                  description: Optional timestamp override when marking the roll finished.
      responses:
        "200":
          description: Roll marked finished
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roll"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/film/rolls/{id}/process:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      operationId: processRoll
      summary: Mark a roll processed
      description: Marks the roll processed/developed. The processed timestamp defaults to now when omitted.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                processed_at:
                  type:
                    - string
                    - "null"
                  format: date-time
                  description: Optional processing timestamp override.
                developed_at:
                  type:
                    - string
                    - "null"
                  format: date-time
                  deprecated: true
                  description: Compatibility alias for `processed_at`.
                development_profile_id:
                  type:
                    - string
                    - "null"
                  description: Optional development profile to attach when processing the roll. The profile must belong to the current user and match the roll's film stock when the roll has a film stock.
                development_notes:
                  type:
                    - string
                    - "null"
                  description: Optional notes recorded when the roll is processed.
      responses:
        "200":
          description: Roll processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roll"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/film/rolls/{id}/reopen:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      operationId: reopenRoll
      summary: Reopen a roll
      description: Reopens a finished or processed roll without erasing processing history fields.
      responses:
        "200":
          description: Roll reopened
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Roll"
        "404":
          $ref: "#/components/responses/NotFound"

  # ── Photographs ───────────────────────────────────────────────────────────────
...